+++ prim_binary.go
@@ -29,7 +29,7 @@ func BinaryAndImpl(args *Data, env *SymbolTableFrame) (result *Data, err error) 
         err = errors.New(fmt.Sprintf("Integer expected, received %s %s", TypeName(TypeOf(arg1)), String(arg1)))
         return
     }
-    b1 := uint32(IntegerValue(arg1))
+    b1 := uint64(IntegerValue(arg1))
 
     arg2, err := Eval(Second(args), env)
     if err != nil {
@@ -39,9 +39,9 @@ func BinaryAndImpl(args *Data, env *SymbolTableFrame) (result *Data, err error) 
         err = errors.New(fmt.Sprintf("Integer expected, received %s %s", TypeName(TypeOf(arg2)), String(arg2)))
         return
     }
-    b2 := uint32(IntegerValue(arg2))
+    b2 := uint64(IntegerValue(arg2))
     
-    return IntegerWithValue(int32(b1 & b2)), nil
+    return IntegerWithValue(int64(b1 & b2)), nil
 }
 
 func BinaryOrImpl(args *Data, env *SymbolTableFrame) (result *Data, err error) {
@@ -53,7 +53,7 @@ func BinaryOrImpl(args *Data, env *SymbolTableFrame) (result *Data, err error) {
         err = errors.New(fmt.Sprintf("Integer expected, received %s %s", TypeName(TypeOf(arg1)), String(arg1)))
         return
     }
-    b1 := uint32(IntegerValue(arg1))
+    b1 := uint64(IntegerValue(arg1))
 
     arg2, err := Eval(Second(args), env)
     if err != nil {
@@ -63,9 +63,9 @@ func BinaryOrImpl(args *Data, env *SymbolTableFrame) (result *Data, err error) {
         err = errors.New(fmt.Sprintf("Integer expected, received %s %s", TypeName(TypeOf(arg2)), String(arg2)))
         return
     }
-    b2 := uint32(IntegerValue(arg2))
+    b2 := uint64(IntegerValue(arg2))
     
-    return IntegerWithValue(int32(b1 | b2)), nil
+    return IntegerWithValue(int64(b1 | b2)), nil
 }
 
 func BinaryNotImpl(args *Data, env *SymbolTableFrame) (result *Data, err error) {
@@ -77,9 +77,9 @@ func BinaryNotImpl(args *Data, env *SymbolTableFrame) (result *Data, err error) 
         err = errors.New(fmt.Sprintf("Integer expected, received %s %s", TypeName(TypeOf(arg1)), String(arg1)))
         return
     }
-    b1 := uint32(IntegerValue(arg1))
+    b1 := uint64(IntegerValue(arg1))
     
-    return IntegerWithValue(int32(b1 ^ uint32(0xFFFFFFFF))), nil
+    return IntegerWithValue(int64(b1 ^ uint64(0xFFFFFFFF))), nil
 }
 
 func LeftShiftImpl(args *Data, env *SymbolTableFrame) (result *Data, err error) {
@@ -91,7 +91,7 @@ func LeftShiftImpl(args *Data, env *SymbolTableFrame) (result *Data, err error) 
         err = errors.New(fmt.Sprintf("Integer expected, received %s %s", TypeName(TypeOf(arg1)), String(arg1)))
         return
     }
-    b1 := uint32(IntegerValue(arg1))
+    b1 := uint64(IntegerValue(arg1))
 
     arg2, err := Eval(Second(args), env)
     if err != nil {
@@ -101,9 +101,9 @@ func LeftShiftImpl(args *Data, env *SymbolTableFrame) (result *Data, err error) 
         err = errors.New(fmt.Sprintf("Integer expected, received %s %s", TypeName(TypeOf(arg2)), String(arg2)))
         return
     }
-    b2 := uint32(IntegerValue(arg2))
+    b2 := uint64(IntegerValue(arg2))
     
-    return IntegerWithValue(int32(b1 << b2)), nil
+    return IntegerWithValue(int64(b1 << b2)), nil
 }
 
 func RightShiftImpl(args *Data, env *SymbolTableFrame) (result *Data, err error) {
@@ -115,7 +115,7 @@ func RightShiftImpl(args *Data, env *SymbolTableFrame) (result *Data, err error)
         err = errors.New(fmt.Sprintf("Integer expected, received %s %s", TypeName(TypeOf(arg1)), String(arg1)))
         return
     }
-    b1 := uint32(IntegerValue(arg1))
+    b1 := uint64(IntegerValue(arg1))
 
     arg2, err := Eval(Second(args), env)
     if err != nil {
@@ -125,7 +125,7 @@ func RightShiftImpl(args *Data, env *SymbolTableFrame) (result *Data, err error)
         err = errors.New(fmt.Sprintf("Integer expected, received %s %s", TypeName(TypeOf(arg2)), String(arg2)))
         return
     }
-    b2 := uint32(IntegerValue(arg2))
+    b2 := uint64(IntegerValue(arg2))
     
-    return IntegerWithValue(int32(b1 >> b2)), nil
+    return IntegerWithValue(int64(b1 >> b2)), nil
 }
